+++ noodles-cram/src/reader/record.rs
@@ -9,6 +9,7 @@ use std::{
 
 use byteorder::ReadBytesExt;
 use noodles_bam as bam;
+use noodles_core::Position;
 use noodles_sam::{
     self as sam,
     record::{quality_scores::Score, sequence::Base},
@@ -523,7 +524,7 @@ where
 
         for _ in 0..feature_count {
             let feature = self.read_feature(prev_position)?;
-            prev_position = feature.position();
+            prev_position = usize::from(feature.position());
             record.add_feature(feature);
         }
 
@@ -559,13 +560,14 @@ where
         .and_then(|n| usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))
     }
 
-    fn read_feature(&mut self, prev_position: i32) -> io::Result<Feature> {
+    fn read_feature(&mut self, prev_position: usize) -> io::Result<Feature> {
         use feature::Code;
 
         let code = self.read_feature_code()?;
 
         let delta = self.read_feature_position()?;
-        let position = prev_position + delta;
+        let position = Position::try_from(prev_position + delta)
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
         match code {
             Code::Bases => {
@@ -642,8 +644,9 @@ where
         })
     }
 
-    fn read_feature_position(&mut self) -> io::Result<i32> {
-        self.compression_header
+    fn read_feature_position(&mut self) -> io::Result<usize> {
+        let encoding = self
+            .compression_header
             .data_series_encoding_map()
             .in_read_positions_encoding()
             .ok_or_else(|| {
@@ -651,14 +654,14 @@ where
                     io::ErrorKind::InvalidData,
                     ReadRecordError::MissingDataSeriesEncoding(DataSeries::InReadPositions),
                 )
-            })
-            .and_then(|encoding| {
-                decode_itf8(
-                    encoding,
-                    &mut self.core_data_reader,
-                    &mut self.external_data_readers,
-                )
-            })
+            })?;
+
+        decode_itf8(
+            encoding,
+            &mut self.core_data_reader,
+            &mut self.external_data_readers,
+        )
+        .and_then(|n| usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))
     }
 
     fn read_stretches_of_bases(&mut self) -> io::Result<Vec<Base>> {